fix installer issue, rename "Base" to "src"
