+++ .gitignore
@@ -1,8 +1,8 @@
 .idea/
 __pycache__/
-Base/apiKey.py
+src/apiKey.py
 song.txt
-test_bit.py
+*test_bit*
 
 # PyBuilder
 target/
